Move AppService to single recorded test class
