--- tools/Azure.Mcp.Tools.AppService/tests/Azure.Mcp.Tools.AppService.Tests/BaseAppServiceCommandLiveTests.cs
@@ -1,51 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System.Text.Json;
-using System.Text.Json.Serialization.Metadata;
-using Microsoft.Mcp.Tests.Client;
-using Microsoft.Mcp.Tests.Client.Helpers;
-using Microsoft.Mcp.Tests.Generated.Models;
-using Xunit;
-
-namespace Azure.Mcp.Tools.AppService.Tests;
-
-public abstract class BaseAppServiceCommandLiveTests(ITestOutputHelper output, TestProxyFixture fixture, LiveServerFixture liveServerFixture)
-    : RecordedCommandTestsBase(output, fixture, liveServerFixture)
-{
-    public override List<BodyKeySanitizer> BodyKeySanitizers =>
-    [
-        ..base.BodyKeySanitizers,
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.selfLink")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.customDomainVerificationId")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.inboundIpAddress")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.possibleInboundIpAddresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.inboundIpv6Address")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.possibleInboundIpv6Addresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.ftpsHostName")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.outboundIpAddresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.possibleOutboundIpAddresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.outboundIpv6Addresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.possibleOutboundIpv6Addresses")),
-        new BodyKeySanitizer(new BodyKeySanitizerBody("$.properties.homeStamp")),
-    ];
-
-    // Sanitize resource group name in response bodies (complements the URI sanitizer in base class)
-    public override List<BodyRegexSanitizer> BodyRegexSanitizers =>
-    [
-        ..base.BodyRegexSanitizers,
-        new BodyRegexSanitizer(new BodyRegexSanitizerBody()
-        {
-            Regex = Settings.ResourceGroupName,
-            Value = "Sanitized"
-        }),
-    ];
-
-    protected static T DeserializeResult<T>(JsonElement? element, JsonTypeInfo<T> jsonTypeInfo)
-    {
-        Assert.NotNull(element);
-        var deserialized = JsonSerializer.Deserialize(element.Value, jsonTypeInfo);
-        Assert.NotNull(deserialized);
-        return deserialized;
-    }
-}
