--- tools/Azure.Mcp.Tools.AppService/tests/Azure.Mcp.Tools.AppService.Tests/Webapp/Diagnostic/DetectorListCommandLiveTests.cs
@@ -1,35 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Azure.Mcp.Tools.AppService.Commands;
-using Microsoft.Mcp.Tests.Client;
-using Microsoft.Mcp.Tests.Client.Helpers;
-using Microsoft.Mcp.Tests.Helpers;
-using Xunit;
-
-namespace Azure.Mcp.Tools.AppService.Tests.Webapp.Diagnostic;
-
-[Trait("Command", "DetectorListCommand")]
-public class DetectorListCommandLiveTests(ITestOutputHelper output, TestProxyFixture fixture, LiveServerFixture liveServerFixture)
-    : BaseAppServiceCommandLiveTests(output, fixture, liveServerFixture)
-{
-    [Fact]
-    public async Task ExecuteAsync_DetectorsList_ReturnsDetectors()
-    {
-        var webappName = RegisterOrRetrieveDeploymentOutputVariable("webappName", "WEBAPPNAME");
-        webappName = TestMode == TestMode.Playback ? "Sanitized-webapp" : webappName;
-        var resourceGroupName = RegisterOrRetrieveVariable("resourceGroupName", Settings.ResourceGroupName);
-
-        var result = await CallToolAsync(
-            "appservice_webapp_diagnostic_list",
-            new()
-            {
-                { "subscription", Settings.SubscriptionId },
-                { "resource-group", resourceGroupName },
-                { "app", webappName }
-            });
-
-        var detectorsResult = DeserializeResult(result, AppServiceJsonContext.Default.DetectorListResult);
-        Assert.NotEmpty(detectorsResult.Detectors);
-    }
-}
