--- tools/Azure.Mcp.Tools.AppService/tests/Azure.Mcp.Tools.AppService.Tests/Webapp/Settings/AppSettingsGetCommandLiveTests.cs
@@ -1,35 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Azure.Mcp.Tools.AppService.Commands;
-using Microsoft.Mcp.Tests.Client;
-using Microsoft.Mcp.Tests.Client.Helpers;
-using Microsoft.Mcp.Tests.Helpers;
-using Xunit;
-
-namespace Azure.Mcp.Tools.AppService.Tests.Webapp.Settings;
-
-[Trait("Command", "AppSettingsGetCommand")]
-public class AppSettingsGetCommandLiveTests(ITestOutputHelper output, TestProxyFixture fixture, LiveServerFixture liveServerFixture)
-    : BaseAppServiceCommandLiveTests(output, fixture, liveServerFixture)
-{
-    [Fact(Skip = "Test temporarily disabled - recording can't consent to secret elicitation")]
-    public async Task ExecuteAsync_AppSettingsList_ReturnsAppSettings()
-    {
-        var webappName = RegisterOrRetrieveDeploymentOutputVariable("webappName", "WEBAPPNAME");
-        webappName = TestMode == TestMode.Playback ? "Sanitized-webapp" : webappName;
-        var resourceGroupName = RegisterOrRetrieveVariable("resourceGroupName", Settings.ResourceGroupName);
-
-        var result = await CallToolAsync(
-            "appservice_webapp_settings_get-appsettings",
-            new()
-            {
-                { "subscription", Settings.SubscriptionId },
-                { "resource-group", resourceGroupName },
-                { "app", webappName }
-            });
-
-        var getResult = DeserializeResult(result, AppServiceJsonContext.Default.AppSettingsGetResult);
-        Assert.NotEmpty(getResult.AppSettings);
-    }
-}
