+++ tools/Azure.Mcp.Tools.AppService/tests/Azure.Mcp.Tools.AppService.Tests/assets.json
@@ -2,5 +2,5 @@
     "AssetsRepo": "Azure/azure-sdk-assets",
     "AssetsRepoPrefixPath": "",
     "TagPrefix": "Azure.Mcp.Tools.AppService.Tests",
-    "Tag": "Azure.Mcp.Tools.AppService.Tests_4ab462af65"
+    "Tag": "Azure.Mcp.Tools.AppService.Tests_f7db92bdbc"
 }
